Improve check whether a phrase is simplified or traditional Chinese
The improvement is to ignore all non-Han characters when
doing the check.

This is to avoid classifying a simplified Chinese string as
traditional just because it happens to include some non-Chinese
characters, for example box drawing characters, which cannot be
converted to gb2312 but happen to be convertible to big5hkscs.

This fixes the problem in the emoji-table input method that most
phrases cannot be input at all.

See: https://bugzilla.redhat.com/show_bug.cgi?id=856320